+++ src/dialogs/MaximaNotStartingDialog.cpp
@@ -22,6 +22,9 @@
 //  SPDX-License-Identifier: GPL-2.0+
 
 #include "MaximaNotStartingDialog.h"
+#include <wx/stattext.h>
+#include <wx/log.h>
+#include "../WrappingStaticText.h"
 
 MaximaNotStartingDialog::MaximaNotStartingDialog(wxWindow *parent, int id, Configuration *config,
                                                  wxString text,
@@ -30,7 +33,11 @@ MaximaNotStartingDialog::MaximaNotStartingDialog(wxWindow *parent, int id, Confi
   : wxDialog(parent, id, text, pos, size,
              style),
     m_configuration(config) {
+  wxLogMessage(_("Asking the user for the location of a working maxima binary"));
   wxBoxSizer *vsizer = new wxBoxSizer(wxVERTICAL);
+  vsizer->Add(new WrappingStaticText(this, wxID_ANY,
+                               _("Maxima, the program that does the actual maths, cannot be started. This might mean that it isn't installed (it can be found at ↗https://maxima.sourceforge.net) or that wxMaxima doesn't know where to search for it. In that case please choose the location of a working maxima binary")),
+              wxSizerFlags());
   wxFlexGridSizer *nameSizer = new wxFlexGridSizer(9, 2, 0, 0);
   m_autodetectMaxima =
     new wxRadioButton(this, wxID_ANY, _("Autodetect"),
@@ -52,19 +59,33 @@ MaximaNotStartingDialog::MaximaNotStartingDialog(wxWindow *parent, int id, Confi
   nameSizer->Add(m_maximaUserLocation,
                  wxSizerFlags().Expand().Border(wxUP | wxDOWN, 0));
   vsizer->Add(nameSizer, 1, wxEXPAND | wxLEFT, 0);
+  m_maximaUserLocation->SetValue(m_configuration->MaximaUserLocation());
+  m_autodetectMaxima->SetValue(m_configuration->AutodetectMaxima());
+  m_noAutodetectMaxima->SetValue(!m_configuration->AutodetectMaxima());
 
   wxBoxSizer *buttonSizer = new wxBoxSizer(wxHORIZONTAL);
 #if defined __WXMSW__
   button_1 = new wxButton(this, wxID_OK, _("OK"));
+  button_1->Connect(wxEVT_BUTTON, wxCommandEventHandler(MaximaNotStartingDialog::OnOkPress),
+                    NULL, this);
   button_1->SetDefault();
   button_2 = new wxButton(this, wxID_CANCEL, _("Cancel"));
 #else
   button_1 = new wxButton(this, wxID_CANCEL, _("Cancel"));
   button_2 = new wxButton(this, wxID_OK, _("OK"));
+  button_2->Connect(wxEVT_BUTTON, wxCommandEventHandler(MaximaNotStartingDialog::OnOkPress),
+                    NULL, this);
   button_2->SetDefault();
 #endif
   buttonSizer->Add(button_1, 0, wxALL, 5);
   buttonSizer->Add(button_2, 0, wxALL, 5);
   vsizer->Add(buttonSizer, 1, wxEXPAND | wxLEFT, 0);
   SetSizerAndFit(vsizer);
+}
+
+void MaximaNotStartingDialog::OnOkPress(wxCommandEvent &event)
+{
+  event.Skip();
+  m_configuration->MaximaUserLocation(m_maximaUserLocation->GetValue());
+  m_configuration->AutodetectMaxima(m_autodetectMaxima->GetValue());
 }